move our envoy type wrappers to types package
