+++ envoy/types/cors_test.go
@@ -1,4 +1,4 @@
-package envoy
+package types
 
 import (
 	"reflect"
